requirements.txt
blurhash==1.1.4
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
decorator==5.1.1
google-api-core==2.11.0
google-api-python-client==2.80.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
googleapis-common-protos==1.58.0
httplib2==0.21.0
idna==3.4
Mastodon.py==1.8.0
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
python-magic==0.4.27
pytz==2022.7.1
requests==2.28.2
rsa==4.9
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.15
